requirements.txt
python-telegram-bot==22.0
openai>=1.0.0
python-dotenv==1.0.1